refactor: keep only one Rest endpoint per resource.
fixes #26
