+++ servicebox-app/src/main/resources/templates/claimForm.html
@@ -41,7 +41,7 @@
          <div id="response"/>
        </div>
        <hr>
-       <a href="/claims">Back</a>
+       <a href="/claim/all">Back</a>
      </div>
     <hr>
   {/body}
