+++ servicebox-app/src/main/resources/templates/home.html
@@ -7,10 +7,10 @@
       <div class="container">
         <div class="row gy-5">
           <div class="col-2">
-            <div class="p-4 btn btn-primary"><a href="/claims" class="text-white">Claims</a></div>
+            <div class="p-4 btn btn-primary"><a href="/claim/all" class="text-white">Claims</a></div>
           </div>
           <div class="col-2">
-            <div class="p-4 btn btn-primary" ><a href="/services" class="text-white">Services</a></div>
+            <div class="p-4 btn btn-primary" ><a href="/service/all" class="text-white">Services</a></div>
           </div>
         </div>
       </div>
