+++ servicebox-app/src/main/resources/templates/serviceForm.html
@@ -41,7 +41,7 @@
          <div id="response"/>
        </div>
        <hr>
-       <a href="/services">Back</a>
+       <a href="/service/all">Back</a>
      </div>
     <hr>
   {/body}
